Refactored the apio scons code to have architecture plugin that provides architecture specific logic (a few scons builders) while having everything else shared (arg parsing, common builders, target definisions). This makes the architecture support intuitive and extensible. No change in functionality.
